feat: bring back FlakinessReport.validate method
It turned out to be a hassle to use Schema directly in clients: there
are zod version mismatches, and we potentially want to have more checks
in the report validation as well.

So this patch drops `FlakinessSchema` and brings back the
`FlakinessReport.validate` method. It also is more clever now: it'll
dump only 5 first issues, and then bail.